requirements.txt
click>=8.0
click-shell>=2.0
python-dotenv>=1.0
requests>=2.28
tabulate>=0.9
rich>=13.0
